Optimize the load performance for large file
