+++ be/src/olap/olap_define.h
@@ -268,6 +268,7 @@ enum OLAPStatus {
     OLAP_ERR_WRITER_DATA_WRITE_ERROR = -1201,
     OLAP_ERR_WRITER_ROW_BLOCK_ERROR = -1202,
     OLAP_ERR_WRITER_SEGMENT_NOT_FINALIZED = -1203,
+    OLAP_ERR_WRITER_FLUSH_QUEUE_SHUTDOWN = -1204,
 
     // RowBlock
     // [-1300, -1400)
